a completely overhauled document with the help of Claude 4:
The attached Specification document structure needs a new structure that helps a potential collaborator or AI to better understand the philosophy of the proposed archutecture.

The introduction is a bit verbose, maybe it can be a bit more condensed.

For the rest, please feel free to reformulate, but make sure its core logic stays the same.If some points need more elaboration from the authors side, please state this too.  if you have suggestion or ideas that would expand on them, please do so, but indicate them in a clear and identifiable way.

---

then it needed some more nudging:

you dropped important information. which I said you should not do under no circumstances. for example, the lazy evaluation algorithm is not complete. or implementation details like the simple data structure of the edge. not all the nodes are defined. etc. I could go on.

I do like the new structure. and maybe it makes sense to add the details into an appendix chapter. Create also a Table of Content and use anchor links to link chapters or concepts that are related to each other.

Look at the original again and make sure all its contents, concepts and ideas find the way into the new document.